フォームタイプとカラムの型変更

+++ db/migrate/20201016060445_create_calculations.rb
@@ -6,17 +6,17 @@ class CreateCalculations < ActiveRecord::Migration[6.0]
 
       t.string :group_name_one
       t.integer :people_one, null: false
-      t.integer :ratio_one, null: false
+      t.float :ratio_one, null: false
       t.integer :price_one, null: false
 
       t.string :group_name_two
       t.integer :people_two, null: false
-      t.integer :ratio_two, null: false
+      t.float :ratio_two, null: false
       t.integer :price_two, null: false
 
       t.string :group_name_three
       t.integer :people_three
-      t.integer :ratio_three
+      t.float :ratio_three
       t.integer :price_three
 
       t.integer :indivisible_number, null: false
